finished cleaning
Recording, writing, and real time recording visualizer are all now
implemented in services/threads.

An additional blocking queue was added to signify the end of file I/O
in order to prevent file access before the writing thread finished.

The WavVisualizer is now a member of the CanvasView, this makes more
sense from an organizational standpoint.

Adding the ability to visualize recording is now as easy as calling
canvas.listenForRecording(current activity). Note that this should only
be done once per active recording.

Loading a waveform to a canvas can be done by
canvas.loadWaveFromFile(absolute file path), then
canvas.displayWaveform(seconds), where seconds is the number of seconds
to display across the canvas. Setting this to 0 will show the entire
waveform in the canvas.