Corrected file writing
Wave files are now written directly (without the use of a RAW file). A
placeholder is written for the header and then filled in upon
completion of recording. This is to prevent having to copy the file
unnecessarily at the end.

TODO:
Analyze and store information for visualization directly rather than
collect all data in an array (that will overflow) and compress after
recording.

Other changes:
removed an unnecessary assignment
removed a debug print statement