Work In Progress
Introduced Memory Mapped IO to cache the file in memory for faster
reading.

This allows for fast reading from disk so the entire file doesn’t have
to be in memory when drawing the waveform.

There is still an issue mapping location to seconds, and thus scrolling
has some stuttering.