Volume Bar
Implemented a volume bar.
The equation should be:
20*Log(value/MAX), but I am multiplying by 10 instead otherwise it
takes a lot for the volume lights to light up. Should calibrate this
with Brad Harrington.

Changed the onCreate to require a user to initiate recording rather
than begin immediately.

The waveform may need to be scaled differently to match up with the
recording lights.

Waveform scrolling still stutters a bit towards the end of playback,
and seems to work differently on different tablets.