Bugfix: Recording Audio
Releases the media player if the user presses the back button. This
caused an issue with playback on future recordings.

Also, placed braces around case statements in exit dialog. Checks are
done to print whether the recording service successfully stopped. Also
onBackPressed now calls onDestroy in hopes that this clears the state
of the activity to prevent issues with restarting recording after the
back button was pressed.

NOTE: I am unable to reproduce the problem now with corrupting the
playback causing future recordings to play back quickly. This used to
be reproducible by pressing back while recording, then recording again.
Keep on a lookout for this behavior.