Bugfix: Save while paused and go back after save
Allows the user to save the recording while playback is paused.

Also correctly goes back when the back button is pressed after saving.
This problem was introduced due to calling on destroy, rather than on
back pressed.

Tried to reproduce the fast recording playback issue, and could not get
it to come up. Braces on case statements may have been the issue?