Bugfix- press back during pause
Fixes the bug where future recordings are corrupted if the user presses
back while recording is paused.

The issue came from not cleaning up resources correctly on the back
pressed due to the isRecording flag being false (due to the pause
state).

This also implies there may have been an issue cleaning up the
MediaPlayer object if playback was paused during the back press, but
this commit handles that case.