BugFix: Pressing back hanging
Fixed an issue where pressing back would cause the app to hang. This was
due to trying to take a message from a queue that wasn't issued.

Ultimately, what needed to happen was to introduce a way to distinguish
between pausing to recording (which needs recording resources cleaned up
on a back press) and pausing to playback (which needs playback resources
cleaned up).

In addition, more condition checks were introduced in the playback class
to stop issues related to operations on a null mediaplayer.