Minimap back, playback bug found
Reimplemented the minimap.

Found a bug in the playback code. I was creating a new thread every time
play is pressed, even if one is currently running. Causes stuttering as
threads compete to draw on the screen (and they will be drawing at many
different locations). I'm not sure how this was ever working. Need to
rework this thread and method, and also keep in mind how to actually
find where you are in playback, as getLocation is not working well.