Restored old settings page
When playing with the settings page and trying to fix how long it takes
to sync, I must have done something that broke it. I have restored an
old commit.

Added code to print how many threads are active in RecordingScreen as a
sanity check.

Minor style changes to WavVisualizer.