Housekeeping, Restored Playback Section
Tried to move the gesture detector to the minimap view from the
recording screen, but it would no longer correctly detect scrolling.

Cleaned up more of RecordingScreen and restored the ability to stop
playback with a selected section. This detection happens on a waveform
canvas draw- I put this here instead of in canvas view's redraw because
I don't want it executed twice.

Renamed the model folder, I think it was just a stock name, but it had
no meaning to me as to what it was.