Load an Existing Audio File
Loads an existing audio file for playback.

Note: this currently has eliminated the use of the compressed audio file
(it's use is commented out both in WavFileLoader and WavVisualizer) Need
to properly handle loading the correct visualization file, cleanup the
files when they are supposed to be deleted, and handle loading without
one (and generate one in a background thread).