Save Visualization Files
Now saving all visualization files.

Loading files will try to load a matching visualization file and
immediately use it for playback. In the absence of a visualization file,
a thread will spawn to create one while falling back on the raw audio
file in the mean time. When the loaded file is available, it will swap
on the next draw.

TODO:
Remove visualization files that do not match with files in the default
folder.
Handle the case where a visualization file does not complete so as not
to load a corrupted one.