Fixed start position computation
Fixed the compressed file equation for computing the start position.

Also make sure that increment is even in all files; an odd increment is
going to split up shorts which will produce incorrect behavior.

If there is still an issue in mapping, it is likely in the compression
thread in WavFileLoader.