Various bugfixes
Fixed the issue where clearing the markers would not immediately draw
the change. I wasn't calling updateUI, which needed to be done in the
playbackscreen method for clearing the markers.

Fixed the issue where playback would resume from the start of the file
upon hitting a stop marker, rather than resuming from the start marker.
This was caused in startPlaybackPosition being given time in ms rather
than the location in the file.