Bugfix: Loaded waveform not displaying
Waveform was not drawing due to the scaling in getDataToDraw being taken
from WavFileWriter; which doesn't work if the file was loaded.

TODO: change the scaling to be db based

Removed the locking around draws, shouldn't be necessary when draws only
happen when one completes. if there's an issue, then use synchronized
anyway.