Accessor introduced
Introduced an accessor class that abstracts away the cut section for
drawing. The compressed file is also abstracted away, so drawing is just
done iteratively as if there is no disruption in the indices.

Currently multiple cuts still behave weird, and only the local and
global views are affected. Still need to add support for the markers.