Bugfix: app restarting on home/apps button press
Playback screen wasn't pushing itself onto the activity stack, as it was
set to no history in the manifest. Changing this appears to have enabled
the ability to return to the app if a button was pressed.

Various logging statements added.

Minimap array declared as a member variable in the visualizer rather
than allocating every time.

Removed unnecessary variables from visualizer methods.