Bugfix
Fixed how saving works. Saving to a different name will make a new file
and the old file will be preserved (UUID from recording still cleaned
up) rather than renaming the file (and thus a "overwrite")

Overwriting without a cut just renames the file, but overwriting with a
cut writes to a new file and deletes the old.

Not overwriting with a cut writes to a new file.