Progress dialog on zip files
Now displaying a progress dialog when zipping files. This required
refactoring export as nothing can run on the UI thread while the dialog
is shown. As such, zipping had to be done in a thread, which then had to
call an abstract method to get the location or other potential user
information, before then actually performing the upload (s3 needs no
user input, but directory/apps/sd card do).