Bugfix: playback thread accessing released player
Without joining the thread, there is no guarantee that it is not in the
middle of execution when the player and buffer references are set to
null.

nulling the references has been moved to the thread upon exiting the
loop, requiring another flag to determine if this behavior is desired.

Furthermore, release is now happening in the onDestroy method rather
than on pause. This was causing problems when the activity was hidden
and recreated.

Also check to make sure when writing a cut that the element to access is
not out of bounds of the mapped file.