Update Recording screen with the new layout
Give source audio playback and file info panel a darker background because
it's considered a "secondary" action. It keeps the recording panel
(containing the record button and time elapsed) and the waveform panel the
main focus of the screen.

Volume bar is given 1 solid color that changes based on decibel level. It
also has a much lower hold delay for a more accurate feedback.

Create UnitPicker "widget" that's basically a stripped down version of the
NumberPicker widget. This widget can be used as chapter, chunk, or verse
picker inside the app. More API (public methods) can be added in the
future as needed.