Update icons in playback and recording screen
Unit picker's increment and decrement buttons are changed to plus and
minus icons. This change is reflected in the recording screen. The old
icons were play icon that was repurposed, which polluted the semantic
meaning of the play icon. It was also quite big that it distracts from the
main record button that's at the bottom main panel.

In the playback screen, there are more noteable changes.

Starting from the top, the minimap icon is now changed to a custom
waveform icon instead of the icon for a physical map.

The start and end marker is given dark cyan color like the one used in the
volume bar. This also to avoid having bright colors clashing everywhere on
the screen.

Secondary action buttons on the main panel are made slightly smaller to
emphasize the main actions more.