Resync on Unit and Chapter Activities
added the resync fragment to the unit and chapter activities.

Fixed an issue in the unit card that was starting a recording intent
with the adapter position for the chapter- which is correct for a
chapter card, but with a unit card it means using the unit as the
chapter which is very wrong.