Closing streams in finally
Closing the various file streams in WavFile.java and WavFileWriter.java

The bug report was caused by a null pointer exception, that appears to
ultimately be as a result of a wav file not being recorded due to the OS
not being able to open it during the recording phase.

Up to this point, the file should only be interacted with in these two
files, where a finally block as been added to close the streams.

Additional logging statements have been included.