Play/Pause works
Short work even if the user pauses right on the end. Was having issues
of it not catching the listener, and the player having data in the
buffer, but not playing. This appears to be a result of an underflow-
not enough data being written to the buffer after a pause.