Hide Keyboard/menu on source audio activity
Pressing back from the file picker or selecting a tr file results in
re-engaging the search bar and opening the keyboard.

The search bar only makes sense when the language list is open. Opening
the language list and selecting the item sets whether the search menu
should be visible and invalidates the menu forcing a redraw- this keeps
me from having to store a reference to the menu (which is probably a bad
idea).