Jbs bugfix#657 (#728)
* Fix intent

Lingering problem from having changed the package name- the intent sent
throught the preference screen still used the old package

* Hide Keyboard/menu on source audio activity

Pressing back from the file picker or selecting a tr file results in
re-engaging the search bar and opening the keyboard.

The search bar only makes sense when the language list is open. Opening
the language list and selecting the item sets whether the search menu
should be visible and invalidates the menu forcing a redraw- this keeps
me from having to store a reference to the menu (which is probably a bad
idea).

* Close keyboard on Settings page

Keyboard should now close when the source language is selected from the
settings page.

Back button fixed (wasn't overriding the correct back button)

Hiding the search menu item similar to the last commit. Instanceof check
seems to be necessary, the fragment is sometimes not correct?