corrected check for section markers
Was checking if the mark and limit had changed; but this check was wrong
if the limit was set to the end of the file.

Now checking with the markerholder if it was actually set.

Also, re-posting the call to swap in the visualization file via a
handler, because the call can theoretically happen before the visualizer
has been created from the VTO callback.