Dev (#797)
* wip

* need to test more thoroughly

Should work,
Should also push off the resync to only be at the chapter level, genesis
at near completion takes a while to resync

* should fix the issue

Threads from the thread pool were not deallocating - needed to set a
timeout

DrawThread was starting twice, currently calling finish on the first
before creating the second. Need to find why it's instantiated twice.

* Jbs further improve resync (#763)

* if something is wrong, then resync

Looks if the db does not have a record for something in the file system
before doing a resync at the chapter or project level

always resync the units (should mostly be in sync, and doesn't take a
while to do compared to the other levels)

Still some cleanup to do

* Fixes issue #757

crashed due to trying to access an element in an empty array, added a
check for size in addition to null

* reorganize

* Changed project resync

Project resync now resyncs the entire tR directory if a discrepency is
found in which projects are expected vs available.

* write done at beginning of file before exit (#782)

in order to know that the visualization thread correctly finished, I
write 4 bytes of zeros at the beginning of the file. Then, before
finishing writing the file, I seek back and write 'D' 'O' 'N' 'E' in
bytes.

Loading the visualization file then needs to check the first four bytes.

* Opening empty chapter crashes (#783)

Opening an empty chapter crashes because of the resync trying to get the
chapterId of a chapter that does not exist.

* corrected check for section markers (#793)

Was checking if the mark and limit had changed; but this check was wrong
if the limit was set to the end of the file.

Now checking with the markerholder if it was actually set.

Also, re-posting the call to swap in the visualization file via a
handler, because the call can theoretically happen before the visualizer
has been created from the VTO callback.

* Stop audio playback when leaving playback activity

audio would continue playing after leaving the playback screen; this
will pause it when the activity finishes

* Jbs bugfix scroll time (#795)

* updating time on scroll

scrolling the minimap was not calling an update to the fragments
containing a timer.

* Elapsed time maintained on mode switch

Elapsed time was being set to zero when fragments swap; now they are
initialized correctly to the current location.

* Bugfix: sort markers

wasn't using the sorted list to write markers to the file

* jan25b

* Bugfix: Adding the same project

Prevents the user from adding the same project; Issues can come up if
this is done but with a different recording mode (chunk vs verse)

* v1.0.0 Beta