fix detection of recording mode (#809)
* fix detection of recording mode

some chunks only have one verse, in which that particular filename cannot determine whether or not the file is a part of a chunk or verse project. Now checking metadata for this information.

* bugfix

intended parameter order to be different

* other issue with parameter order

adding take after a cut also had the parameters mixed up.