Small fixes
Checking if start position + increment is within the accessor's size
results in cutting off early if the capacity is inside this range.
Results in the line SOMETIMES showing up and SOMETIMES not

Also, a rounding error can prevent skipping the last index of a cut,
which can result in rendering values that should be inside of a cut.