Jbs line at end (#849)
* Small fixes

Checking if start position + increment is within the accessor's size
results in cutting off early if the capacity is inside this range.
Results in the line SOMETIMES showing up and SOMETIMES not

Also, a rounding error can prevent skipping the last index of a cut,
which can result in rendering values that should be inside of a cut.

* build warning fix

fixes the compiler warnings due to zip4j

some dex stuff was recommended and took care of a few of the issues, but
the proguard rules turned out to fix everything so the dex stuff was
commended out. links to the stack overflow pages are in the comments

* mapping error

Click at the end related to mapping some of the metadata. Length
parameter for the file map is not the index to stop at, but the length
from the offset provided; so adding the header length to the end is only
going to map the first 44 bytes of metadata

* rename stream variables

shouldn't have been fis to begin with, temp changed to bos

* wip

* wip

* wip

* should fix concurrent access exception

cut operations that remove multiple verse markers had previously thrown
a concurrent access exception; had to do with iterating over the
collection and removing, rather than creating a list to loop over.

* insert and mapped visualization fix

insert needed to switch to relative frame (because a cut will be written
before the insert takes place) and was originally absolute time.

visualization file was accidentally swapping in the raw audio file and
not the visualization file.

* cleanup

* psa fix for database resync (#848)

getParentDirectory was not getting the 3 digit chapter folder required
for the book of psalms

project replaced by  anthology for consistency

* cleanup

minimap now uses frames rather than time, uses the handler to submit the
call to switching from uncompressed to compressed (so as not to happen
between draw calls), makes sure that the minimap invalidation is
synchronized and cannot happen between draws

* 1.0.5-test4