Dev (#858)
* wip

* need to test more thoroughly

Should work,
Should also push off the resync to only be at the chapter level, genesis
at near completion takes a while to resync

* should fix the issue

Threads from the thread pool were not deallocating - needed to set a
timeout

DrawThread was starting twice, currently calling finish on the first
before creating the second. Need to find why it's instantiated twice.

* Jbs further improve resync (#763)

* if something is wrong, then resync

Looks if the db does not have a record for something in the file system
before doing a resync at the chapter or project level

always resync the units (should mostly be in sync, and doesn't take a
while to do compared to the other levels)

Still some cleanup to do

* Fixes issue #757

crashed due to trying to access an element in an empty array, added a
check for size in addition to null

* reorganize

* Changed project resync

Project resync now resyncs the entire tR directory if a discrepency is
found in which projects are expected vs available.

* write done at beginning of file before exit (#782)

in order to know that the visualization thread correctly finished, I
write 4 bytes of zeros at the beginning of the file. Then, before
finishing writing the file, I seek back and write 'D' 'O' 'N' 'E' in
bytes.

Loading the visualization file then needs to check the first four bytes.

* Opening empty chapter crashes (#783)

Opening an empty chapter crashes because of the resync trying to get the
chapterId of a chapter that does not exist.

* corrected check for section markers (#793)

Was checking if the mark and limit had changed; but this check was wrong
if the limit was set to the end of the file.

Now checking with the markerholder if it was actually set.

Also, re-posting the call to swap in the visualization file via a
handler, because the call can theoretically happen before the visualizer
has been created from the VTO callback.

* Stop audio playback when leaving playback activity

audio would continue playing after leaving the playback screen; this
will pause it when the activity finishes

* Jbs bugfix scroll time (#795)

* updating time on scroll

scrolling the minimap was not calling an update to the fragments
containing a timer.

* Elapsed time maintained on mode switch

Elapsed time was being set to zero when fragments swap; now they are
initialized correctly to the current location.

* Bugfix: sort markers

wasn't using the sorted list to write markers to the file

* jan25b

* Bugfix: Adding the same project

Prevents the user from adding the same project; Issues can come up if
this is done but with a different recording mode (chunk vs verse)

* v1.0.0 Beta

* Bugfix: chapter resync crashing on invalid names

chapter folders that do not parse to integers cause crashes, this fix
skips over them.

Also fixed a lingering issue from swapping the end verse from -1 for
verses/chunks to the start verse.

* update version

* fix detection of recording mode (#809)

* fix detection of recording mode

some chunks only have one verse, in which that particular filename cannot determine whether or not the file is a part of a chunk or verse project. Now checking metadata for this information.

* bugfix

intended parameter order to be different

* other issue with parameter order

adding take after a cut also had the parameters mixed up.

* fixed marker behavior on cut (#813)

Markers in a cut are removed

Otherwise, markers are remapped to the correct relative location after
cuts and undo.

* Bugfix: parsing metadata loop

if a limit is set on a buffer, the position will never hit the capacity,
and the while loop will never exit

* Bugfix #815

placing a start marker and moving it without having placed an end marker
will attempt to get the X position of a non-existent end marker.

Solved by checking if the marker exists first

* Jbs line at end (#849)

* Small fixes

Checking if start position + increment is within the accessor's size
results in cutting off early if the capacity is inside this range.
Results in the line SOMETIMES showing up and SOMETIMES not

Also, a rounding error can prevent skipping the last index of a cut,
which can result in rendering values that should be inside of a cut.

* build warning fix

fixes the compiler warnings due to zip4j

some dex stuff was recommended and took care of a few of the issues, but
the proguard rules turned out to fix everything so the dex stuff was
commended out. links to the stack overflow pages are in the comments

* mapping error

Click at the end related to mapping some of the metadata. Length
parameter for the file map is not the index to stop at, but the length
from the offset provided; so adding the header length to the end is only
going to map the first 44 bytes of metadata

* rename stream variables

shouldn't have been fis to begin with, temp changed to bos

* wip

* wip

* wip

* should fix concurrent access exception

cut operations that remove multiple verse markers had previously thrown
a concurrent access exception; had to do with iterating over the
collection and removing, rather than creating a list to loop over.

* insert and mapped visualization fix

insert needed to switch to relative frame (because a cut will be written
before the insert takes place) and was originally absolute time.

visualization file was accidentally swapping in the raw audio file and
not the visualization file.

* cleanup

* psa fix for database resync (#848)

getParentDirectory was not getting the 3 digit chapter folder required
for the book of psalms

project replaced by  anthology for consistency

* cleanup

minimap now uses frames rather than time, uses the handler to submit the
call to switching from uncompressed to compressed (so as not to happen
between draw calls), makes sure that the minimap invalidation is
synchronized and cannot happen between draws

* 1.0.5-test4

* wip to check if rounding helped (#857)

* 1.0.5 beta