Bugfix: Ringbuffer out of bounds
Not sure how the compression array was getting out of bounds, but I made
it so it (and the ring buffer) are created inside the thread so that in
the event that there are multiple threads spawned somehow, they aren't
trying to write at the same time. Also moved the passing of data off to
the ring buffer before accessing the array index, though I still feel
like this bounds checking shouldn't have been an issue?

Try/catch for importing a wav file- if parsing throws exceptions, the
log will now report the name of the file. TODO: give user feedback here-
allow them to delete the file or try to salvage it maybe?

Also, needed to set the language and book view on the main menu screen
if the project id doesn't exist.