Remove anthology from slug structs
anthology needs extra context rather than just regex to be determined,
thus it was removed.

Util functions to get the name without take don't need to reconstruct
the filename, just regex off the end, so that was refactored.

Still providing one method to reconstruct the filename however, but it
should only be temporary as it won't scale with arbitrary plugins aside
from bible and obs.