Bugfix: unit picker
when opening the recording screen and changing the unit or chapter, the
internal state of the file bar fragment is updated, but not the
activity. As a result, the internal values for the activity never
change, and pass the wrong values to the playback activity.

now asking for the value from the fragment instead of using the member
variable